Prisma impact on other components + ./utils directory renamed to ./libs
