+++ libs/KeyboardShortcutHandler.ts
@@ -5,6 +5,9 @@ export const registerShortcuts = (shortcutMap: ShortcutMap) => {
     const handleKeyPress = (event: KeyboardEvent) => {
         const keyCombination = getShortcutCombination(event);
 
+        if(!keyCombination) return;
+
+
         const handler = shortcutMap[keyCombination];
         if (handler) {
             handler(event);
@@ -19,12 +22,16 @@ export const registerShortcuts = (shortcutMap: ShortcutMap) => {
     };
 };
 
-const getShortcutCombination = (event: KeyboardEvent): string => {
+const getShortcutCombination = (event: KeyboardEvent): string|null => {
+    if(!event.key) return null;
+
+
     let keys = [];
 
     if (event.ctrlKey) keys.push("ctrl");
     if (event.altKey) keys.push("alt");
     if (event.shiftKey) keys.push("shift");
+
 
     keys.push(event.key.toLowerCase());
