--- models/Mod.ts
@@ -1,5 +1,0 @@
-interface Mod {
-    mod_id: string
-    displayName: string
-    description: string
-}
