--- models/Player.ts
@@ -1,4 +1,0 @@
-interface Player {
-    player_uid: string
-    player_name: string
-}
